+++ src/publisher/react-integration/dispatcher-trap.ts
@@ -1,7 +1,16 @@
-import { ReactDispatcherTrapApi, ReactInternals } from "../types";
+import {
+  Fiber,
+  ReactContext,
+  ReactDispatcherTrapApi,
+  ReactInternals,
+} from "../types";
 
 type Dispatcher = any;
 type DispatchFn = (state: any) => any;
+type UseContextPathMap = Map<
+  ReactContext<any>,
+  Map<number, { path: string[] | undefined }>
+>;
 
 function extractHookPath() {
   const stack = String(new Error().stack).split("\n").slice(4);
@@ -26,11 +35,15 @@ export function dispatcherTrap(
   renderer: ReactInternals
 ): ReactDispatcherTrapApi {
   let currentDispatcher: Dispatcher | null = null;
+  let currentFiber: Fiber | null = null;
+  let currentUseContext: UseContextPathMap | null = null;
+  let useContextCallIndex = 0;
   const knownDispatcher = new Set();
   const patchedHookFn = new WeakMap<
     DispatchFn,
     { fn: DispatchFn; path: string[] | undefined }
   >();
+  const useContextPaths = new WeakMap<Fiber, UseContextPathMap>();
 
   // function patchUseEffect(dispatcher: Dispatcher) {
   //   const orig = dispatcher.useEffect;
@@ -72,7 +85,7 @@ export function dispatcherTrap(
   function patchStateHook(hookName: string, dispatcher: Dispatcher) {
     const orig = dispatcher[hookName];
 
-    dispatcher[hookName] = function (...args: any[]) {
+    dispatcher[hookName] = (...args: any[]) => {
       const [state, dispatch] = orig(...args);
 
       if (!patchedHookFn.has(dispatch)) {
@@ -89,11 +102,48 @@ export function dispatcherTrap(
     };
   }
 
+  function patchUseContextHook(dispatcher: Dispatcher) {
+    const orig = dispatcher.useContext;
+
+    dispatcher.useContext = (context: ReactContext<any>, ...args: any[]) => {
+      const value = orig(context, ...args);
+
+      if (currentUseContext === null) {
+        if (useContextPaths.has(currentFiber as Fiber)) {
+          currentUseContext = useContextPaths.get(
+            currentFiber as Fiber
+          ) as UseContextPathMap;
+        } else {
+          currentUseContext = new Map();
+          useContextPaths.set(currentFiber as Fiber, currentUseContext);
+        }
+      }
+
+      let contextPaths = currentUseContext.get(context);
+
+      if (!contextPaths) {
+        contextPaths = new Map();
+        currentUseContext.set(context, contextPaths);
+      }
+
+      if (!contextPaths.has(useContextCallIndex)) {
+        contextPaths.set(useContextCallIndex, {
+          path: extractHookPath(),
+        });
+      }
+
+      useContextCallIndex++;
+
+      return value;
+    };
+  }
+
   function patchDispatcher(dispatcher: Dispatcher) {
     if (dispatcher && !knownDispatcher.has(dispatcher)) {
       knownDispatcher.add(dispatcher);
       patchStateHook("useState", dispatcher);
       patchStateHook("useReducer", dispatcher);
+      patchUseContextHook(dispatcher);
       // patchUseEffect(dispatcher);
     }
 
@@ -105,6 +155,9 @@ export function dispatcherTrap(
       return currentDispatcher;
     },
     set: value => {
+      currentFiber = renderer.getCurrentFiber();
+      useContextCallIndex = 0;
+      currentUseContext = null;
       currentDispatcher = patchDispatcher(value);
     },
   });
@@ -112,6 +165,10 @@ export function dispatcherTrap(
   return {
     getHookPath(dispatch: DispatchFn) {
       return patchedHookFn.get(dispatch)?.path;
+    },
+    getFiberUseContextPaths(fiber: Fiber, context: ReactContext<any>) {
+      const contextPaths = useContextPaths.get(fiber)?.get(context);
+      return contextPaths ? [...contextPaths.values()] : undefined;
     },
   };
 }